+++ vignettes/v0_plots.R
@@ -46,15 +46,21 @@ ggPedigree(
   personID = "personID",
   config = list(
     code_male = 1, # Here, 1 = male, 0 = female
-    sex_color_include = FALSE,
-    line_width = 1,
+    sex_color_include = TRUE,
+    sex_color_palette = c("pink", "blue"),
+    point_size = 5,
+    outline_multiplier = 1.25,
+    line_width = 3.5,
+    sex_shape_female = "F",
+    sex_shape_male = "z",
     segment_spouse_color = viridis_pal()(5)[1],
     segment_sibling_color = viridis_pal()(5)[2],
     segment_parent_color = viridis_pal()(5)[3],
     segment_offspring_color = viridis_pal()(5)[4],
+    segment_mz_color = viridis_pal()(5)[5],
     segment_linetype = 3,
     outline_include = TRUE,
-    outline_color = viridis_pal()(5)[5]
+    outline_color = "grey" # viridis_pal()(5)[5]
   )
 )
 
@@ -94,7 +100,7 @@ p <- ggPedigree(
     sex_color_include = TRUE,
     status_code_affected = TRUE,
     status_code_unaffected = FALSE,
-    status_affected_shape = 4
+    status_shape_affected = 4
   )
 )
 
@@ -113,7 +119,7 @@ ggPedigree(
     status_code_unaffected = FALSE,
     status_label_affected = "Infected",
     status_label_unaffected = "Not infected",
-    status_affected_legend_title = "Status"
+    status_legend_title = "Status"
   )
 )
 
@@ -133,9 +139,10 @@ ggPedigree(
   status_column = "proband",
   config = list(
     sex_color_include = TRUE,
+    status_include = TRUE,
     status_code_affected = TRUE,
     status_code_unaffected = FALSE,
-    status_affected_shape = 8 # star shape
+    status_shape_affected = 8 # star shape
   )
 )
 
@@ -146,9 +153,9 @@ ggPedigree(potter,
   config = list(
     focal_fill_personID = 7,
     focal_fill_include = TRUE,
-    focal_fill_high_color = "yellow",
-    focal_fill_mid_color = "red",
-    focal_fill_low_color = "#0D082AFF",
+    #  focal_fill_high_color = "yellow",
+    #  focal_fill_mid_color = "red",
+    #   focal_fill_low_color = "#0D082AFF",
     focal_fill_force_zero = TRUE,
     focal_fill_na_value = "black",
     focal_fill_scale_midpoint = 0.5,
@@ -205,6 +212,30 @@ m2 <- ggPedigree(potter,
 library(patchwork) # for combining plots
 m1 + m2 + plot_layout(ncol = 2) +
   plot_annotation(title = "Mitochondrial Relatives of Harry Potter and Ginny Weasley")
+
+## ----focal_fill_ginny---------------------------------------------------------
+test <- ggPedigree(potter,
+  famID = "famID",
+  personID = "personID",
+  config = list(
+    focal_fill_personID = 8,
+    focal_fill_include = TRUE,
+    # focal_fill_mid_color = "white",
+    focal_fill_low_color = "black",
+    focal_fill_scale_midpoint = 0.55,
+    focal_fill_component = "matID",
+    focal_fill_method = "viridis_d",
+    focal_fill_viridis_option = "turbo",
+    focal_fill_n_breaks = 19,
+    focal_fill_legend_show = FALSE,
+    focal_fill_legend_title = "Mitochondrial Relatives",
+    sex_color_include = FALSE,
+    overlay_include = FALSE
+  ) # highlight Harry Potter
+  # config  = list(segment_mz_color = NA) # color for monozygotic twins
+) + ggplot2::guides(shape = "none")
+
+test
 
 ## ----facet_wrap---------------------------------------------------------------
 p +